requirements.txt
streamlit
paho-mqtt
matplotlib
streamlit-autorefresh
